+++ docs/index.jsx
@@ -23,8 +23,8 @@ import Input from '../src/components/Input';
 import InputExample from '../src/components/Input/example';
 import Label from '../src/components/Label';
 import LabelExample from '../src/components/Label/example';
-import OverlayTrigger from '../src/components/OverlayTrigger';
-import OverlayTriggerExample from '../src/components/OverlayTrigger/example';
+import OverlayWrapper from '../src/components/OverlayWrapper';
+import OverlayWrapperExample from '../src/components/OverlayWrapper/example';
 import Popover from '../src/components/Popover';
 import PopoverExample from '../src/components/Popover/example';
 import Table from '../src/components/Table';
@@ -69,8 +69,8 @@ const componentsArray = [
     example: LabelExample,
   },
   {
-    index: OverlayTrigger,
-    example: OverlayTriggerExample,
+    index: OverlayWrapper,
+    example: OverlayWrapperExample,
   },
   {
     index: Popover,